# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kanbalone API
  version: 0.9.18
  summary: Ultra-light local personal kanban API for human and AI collaboration
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:3000
security: []
paths:
...
  /api/meta:
    get:
      summary: Get app metadata
      operationId: getMeta
      responses:
        "200":
          description: App metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  version:
                    type: string
                  remoteProviders:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        hasCredential:
                          type: boolean
                      required: [id, hasCredential]
                required: [name, version, remoteProviders]
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/remote-diagnostics:
    get:
      summary: List remote provider diagnostic status
      operationId: listRemoteProviderDiagnostics
      responses:
        "200":
          description: Remote provider credential configuration status
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        hasCredential:
                          type: boolean
                        status:
                          type: string
                          enum: [configured, missing_credential]
                      required: [id, hasCredential, status]
                required: [providers]
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      summary: Check a remote issue credential
      description: Attempts to resolve a remote issue with an exact-scope configured provider credential and returns a token-safe diagnostic result. Wildcard credentials are not used for user-supplied diagnostic URLs.
      operationId: checkRemoteCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoteDiagnosticRequest"
      responses:
        "200":
          description: Remote issue is reachable with the configured credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteDiagnosticResult"
        "400":
          description: Diagnostic failed or the provider is unsupported/missing credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteDiagnosticResult"
        "500":
          $ref: "#/components/responses/Error"
...
  /api/tickets/{ticketId}:
    get:
      summary: Get ticket detail
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update ticket
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpsert"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete ticket
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
  /api/tickets/{ticketId}/comments:
    get:
      summary: List comments
      description: Returns comments with the newest comment first.
      operationId: listComments
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Comment list
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                required: [comments]
        "404":
          $ref: "#/components/responses/Error"
    post:
      summary: Add comment
      operationId: addComment
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bodyMarkdown:
                  type: string
              required: [bodyMarkdown]
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/comments/{commentId}:
    patch:
      summary: Update comment
      operationId: updateComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bodyMarkdown:
                  type: string
              required: [bodyMarkdown]
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete comment
      operationId: deleteComment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
  /api/comments/{commentId}/push-remote:
    post:
      summary: Push a local comment to the tracked remote issue
      description: Posts the local comment body to the linked remote issue and marks the comment as pushed. Already-pushed and in-progress comments are rejected to avoid duplicate remote comments.
      operationId: pushCommentToRemote
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comment with updated remote sync state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: integer
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            required: [error]
  schemas:
...
    Tag:
      type: object
      properties:
        id: { type: integer }
        boardId: { type: integer }
        name: { type: string }
        color: { type: string }
      required: [id, boardId, name, color]
    CommentSync:
      type: object
      properties:
        commentId: { type: integer }
        status:
          type: string
          enum: [local_only, pushing, pushed, push_failed]
        remoteCommentId:
          type: [string, "null"]
        pushedAt:
          type: [string, "null"]
          format: date-time
        lastError:
          type: [string, "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [commentId, status, remoteCommentId, pushedAt, lastError, createdAt, updatedAt]
    Comment:
      type: object
      properties:
        id: { type: integer }
        ticketId: { type: integer }
        bodyMarkdown: { type: string }
        bodyHtml: { type: string }
        createdAt: { type: string, format: date-time }
        sync:
          $ref: "#/components/schemas/CommentSync"
      required: [id, ticketId, bodyMarkdown, bodyHtml, createdAt, sync]
...
    TicketRelation:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        laneId: { type: integer }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        ref: { type: string }
        shortRef: { type: string }
      required: [id, title, laneId, isResolved, isCompleted, priority, ref, shortRef]
...
    TicketRemote:
      type: object
      properties:
        ticketId: { type: integer }
        provider: { type: string }
        instanceUrl: { type: string }
        resourceType: { type: string }
        projectKey: { type: string }
        issueKey: { type: string }
        displayRef: { type: string }
        url: { type: string }
        title: { type: string }
        bodyMarkdown:
          type: string
          description: Read-only remote issue body snapshot. The ticket bodyMarkdown remains the editable local implementation body.
        bodyHtml:
          type: string
          description: Sanitized HTML rendered from bodyMarkdown.
        state:
          type: [string, "null"]
        remoteUpdatedAt:
          type: [string, "null"]
          format: date-time
        lastSyncedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [ticketId, provider, instanceUrl, resourceType, projectKey, issueKey, displayRef, url, title, bodyMarkdown, bodyHtml, state, remoteUpdatedAt, lastSyncedAt, createdAt, updatedAt]
...
    RemoteDiagnosticRequest:
      type: object
      properties:
        provider:
          type: string
          description: "Remote adapter name. Current providers: github, gitlab, redmine."
        instanceUrl:
          type: string
        projectKey:
          type: string
        issueKey:
          type: string
        url:
          type: string
          description: Remote issue URL to check.
      required: [provider]
    RemoteDiagnosticResult:
      type: object
      properties:
        provider:
          type: string
        hasCredential:
          type: boolean
        status:
          type: string
          enum: [missing_credential, reachable, auth_failed, permission_failed, not_found, rate_limited, unsupported_provider, error]
        displayRef:
          type: string
        url:
          type: string
        message:
          type: string
      required: [provider, hasCredential, status]
    Ticket:
      type: object
      properties:
        id: { type: integer }
        boardId: { type: integer }
        laneId: { type: integer }
        parentTicketId:
          type: [integer, "null"]
        title: { type: string }
        bodyMarkdown: { type: string }
        bodyHtml: { type: string }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        isArchived: { type: boolean }
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        position: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        blockerIds:
          type: array
          items:
            type: integer
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        blockedBy:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        parent:
          oneOf:
            - $ref: "#/components/schemas/TicketRelation"
            - type: "null"
        children:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        ref: { type: string }
        shortRef: { type: string }
        remote:
          oneOf:
            - $ref: "#/components/schemas/TicketRemote"
            - type: "null"
      required:
        [id, boardId, laneId, parentTicketId, title, bodyMarkdown, bodyHtml, isResolved, isCompleted, isArchived, priority, position, createdAt, updatedAt, tags, comments, blockerIds, blockers, blockedBy, parent, children, ref, shortRef, remote]
...
    TicketUpsert:
      type: object
      properties:
        laneId: { type: integer }
        parentTicketId:
          type: [integer, "null"]
        title: { type: string }
        bodyMarkdown: { type: string }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        isArchived: { type: boolean }
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        tagIds:
          type: array
          items:
            type: integer
        blockerIds:
          type: [array, "null"]
          items:
            type: integer
...